Extract out format-independent reader files
Moves format-generic reader files into dwio/common.
Divides struct and floating point column readers into generic and DWRF-specific parts.
