+++ velox/connectors/hive/HiveConnector.cpp
@@ -18,7 +18,6 @@
 #include <memory>
 
 #include "velox/dwio/common/InputStream.h"
-#include "velox/dwio/dwrf/reader/SelectiveColumnReader.h"
 #include "velox/expression/FieldReference.h"
 #include "velox/type/Conversions.h"
 #include "velox/type/Type.h"
