+++ velox/dwio/common/SelectiveColumnReader.cpp
@@ -14,18 +14,15 @@
  * limitations under the License.
  */
 
-#include "velox/dwio/dwrf/reader/SelectiveColumnReaderInternal.h"
+#include "velox/dwio/common/SelectiveColumnReaderInternal.h"
 
-namespace facebook::velox::dwrf {
+namespace facebook::velox::dwio::common {
 
 using dwio::common::TypeWithId;
 using dwio::common::typeutils::CompatChecker;
 
-// Buffer size for reading length stream
-constexpr uint64_t BUFFER_SIZE = 1024;
-
-common::AlwaysTrue& alwaysTrue() {
-  static common::AlwaysTrue alwaysTrue;
+velox::common::AlwaysTrue& alwaysTrue() {
+  static velox::common::AlwaysTrue alwaysTrue;
   return alwaysTrue;
 }
 
@@ -48,7 +45,7 @@ void ScanState::updateRawState() {
 SelectiveColumnReader::SelectiveColumnReader(
     std::shared_ptr<const dwio::common::TypeWithId> requestedType,
     dwio::common::FormatParams& params,
-    common::ScanSpec& scanSpec,
+    velox::common::ScanSpec& scanSpec,
     const TypePtr& type)
     : memoryPool_(params.pool()),
       nodeType_(requestedType),
@@ -302,4 +299,4 @@ void SelectiveColumnReader::resetFilterCaches() {
   }
 }
 
-} // namespace facebook::velox::dwrf
+} // namespace facebook::velox::dwio::common
