+++ velox/dwio/common/SelectiveColumnReaderInternal.h
@@ -19,11 +19,8 @@
 #include "velox/common/base/Portability.h"
 #include "velox/dwio/common/DirectDecoder.h"
 #include "velox/dwio/common/TypeUtils.h"
-#include "velox/dwio/dwrf/common/FloatingPointDecoder.h"
-#include "velox/dwio/dwrf/common/RLEv1.h"
-#include "velox/dwio/dwrf/reader/ColumnVisitors.h"
-#include "velox/dwio/dwrf/reader/SelectiveColumnReader.h"
-#include "velox/dwio/dwrf/utils/ProtoUtils.h"
+#include "velox/dwio/common/ColumnVisitors.h"
+#include "velox/dwio/common/SelectiveColumnReader.h"
 #include "velox/exec/AggregationHook.h"
 #include "velox/type/Timestamp.h"
 #include "velox/vector/ConstantVector.h"
@@ -32,9 +29,9 @@
 
 #include <numeric>
 
-namespace facebook::velox::dwrf {
+namespace facebook::velox::dwio::common {
 
-common::AlwaysTrue& alwaysTrue();
+  velox::common::AlwaysTrue& alwaysTrue();
 
 class Timer {
  public:
@@ -47,16 +44,6 @@ class Timer {
  private:
   const uint64_t startClocks_;
 };
-
-inline RleVersion convertRleVersion(proto::ColumnEncoding_Kind kind) {
-  switch (static_cast<int64_t>(kind)) {
-    case proto::ColumnEncoding_Kind_DIRECT:
-    case proto::ColumnEncoding_Kind_DICTIONARY:
-      return RleVersion_1;
-    default:
-      DWIO_RAISE("Unknown encoding in convertRleVersion");
-  }
-}
 
 template <typename T>
 void SelectiveColumnReader::ensureValuesCapacity(vector_size_t numRows) {
@@ -360,7 +347,5 @@ void SelectiveColumnReader::filterNulls(
   }
   readOffset_ += rows.back() + 1;
 }
-
-std::vector<uint64_t> toPositions(const proto::RowIndexEntry& entry);
 
 } // namespace facebook::velox::dwrf
