+++ velox/dwio/dwrf/reader/SelectiveByteRleColumnReader.h
@@ -16,12 +16,13 @@
 
 #pragma once
 
+#include "velox/dwio/common/SelectiveColumnReaderInternal.h"
 #include "velox/dwio/dwrf/reader/DwrfData.h"
-#include "velox/dwio/dwrf/reader/SelectiveColumnReaderInternal.h"
 
 namespace facebook::velox::dwrf {
 
-class SelectiveByteRleColumnReader : public SelectiveColumnReader {
+class SelectiveByteRleColumnReader
+    : public dwio::common::SelectiveColumnReader {
  public:
   using ValueType = int8_t;
 
@@ -143,7 +144,7 @@ void SelectiveByteRleColumnReader::readHelper(
     ExtractValues extractValues) {
   readWithVisitor(
       rows,
-      ColumnVisitor<int8_t, TFilter, ExtractValues, isDense>(
+      dwio::common::ColumnVisitor<int8_t, TFilter, ExtractValues, isDense>(
           *reinterpret_cast<TFilter*>(filter), this, rows, extractValues));
 }
 
@@ -199,11 +200,15 @@ void SelectiveByteRleColumnReader::processValueHook(
   switch (hook->kind()) {
     case aggregate::AggregationHook::kSumBigintToBigint:
       readHelper<common::AlwaysTrue, isDense>(
-          &alwaysTrue(), rows, ExtractToHook<SumHook<int64_t, int64_t>>(hook));
+          &dwio::common::alwaysTrue(),
+          rows,
+          dwio::common::ExtractToHook<SumHook<int64_t, int64_t>>(hook));
       break;
     default:
       readHelper<common::AlwaysTrue, isDense>(
-          &alwaysTrue(), rows, ExtractToGenericHook(hook));
+          &dwio::common::alwaysTrue(),
+          rows,
+          dwio::common::ExtractToGenericHook(hook));
   }
 }
