+++ velox/dwio/dwrf/reader/SelectiveDwrfReader.h
@@ -16,16 +16,16 @@
 
 #pragma once
 
+#include "velox/dwio/common/SelectiveColumnReader.h"
 #include "velox/dwio/dwrf/reader/ColumnReader.h"
 #include "velox/dwio/dwrf/reader/DwrfData.h"
-#include "velox/dwio/dwrf/reader/SelectiveColumnReader.h"
 
 namespace facebook::velox::dwrf {
 
 // Wrapper for static functions for making DWRF readers
 class SelectiveDwrfReader {
  public:
-  static std::unique_ptr<SelectiveColumnReader> build(
+  static std::unique_ptr<dwio::common::SelectiveColumnReader> build(
       const std::shared_ptr<const dwio::common::TypeWithId>& requestedType,
       const std::shared_ptr<const dwio::common::TypeWithId>& dataType,
       DwrfParams& params,
@@ -33,7 +33,7 @@ class SelectiveDwrfReader {
 
   // Compatibility wrapper for tests. Takes the components of DwrfParams as
   // separate.
-  static std::unique_ptr<SelectiveColumnReader> build(
+  static std::unique_ptr<dwio::common::SelectiveColumnReader> build(
       const std::shared_ptr<const dwio::common::TypeWithId>& requestedType,
       const std::shared_ptr<const dwio::common::TypeWithId>& dataType,
       StripeStreams& stripe,
@@ -50,7 +50,7 @@ class SelectiveColumnReaderFactory : public ColumnReaderFactory {
       std::shared_ptr<common::ScanSpec> scanSpec)
       : scanSpec_(scanSpec) {}
 
-  std::unique_ptr<SelectiveColumnReader> buildSelective(
+  std::unique_ptr<dwio::common::SelectiveColumnReader> buildSelective(
       const std::shared_ptr<const dwio::common::TypeWithId>& requestedType,
       const std::shared_ptr<const dwio::common::TypeWithId>& dataType,
       StripeStreams& stripe,
